Added Move constructor to Vector2 and Vector4 Added lua_Number constructors to Vector2 and Vector4 Finished GameFileImage, GameFileAudio, GameFileObject and GameFileRaw Sprites now contain reference to their texture

+++ scr/Engine/component/Position.cpp
@@ -32,9 +32,7 @@ int Position::lGetPosition(lua_State* L){
 int Position::lSetPosition(lua_State* L){
 	const auto p = Util::luaP_touserdata<Position>(L, 1); 
 	p->setPosition(
-		Math::Vector2(
-		static_cast<int>(lua_tonumber(L, 2)),
-		static_cast<int>(lua_tonumber(L, 3))
-		));
+		Math::Vector2(lua_tonumber(L, 2), lua_tonumber(L, 3))
+		);
 	return 0;
 }
